graduates Label2 to main bundle
