+++ src/deprecated/index.ts
@@ -24,8 +24,8 @@ export type {
 } from '../Dropdown'
 export {default as FormGroup} from '../FormGroup'
 export type {FormGroupProps, FormGroupLabelProps} from '../FormGroup'
-export {default as Label} from '../Label'
-export type {LabelProps} from '../Label'
+export {default as Label} from '../Label_deprecated'
+export type {LabelProps} from '../Label_deprecated'
 export {default as SelectMenu} from '../SelectMenu'
 export type {
   SelectMenuProps,
