+++ src/index.ts
@@ -97,6 +97,8 @@ export type {HeaderProps, HeaderItemProps, HeaderLinkProps} from './Header'
 export {default as Heading} from './Heading'
 export type {HeadingProps} from './Heading'
 export {default as InputField} from './InputField'
+export {default as Label} from './Label'
+export type {LabelProps, LabelColorOptions} from './Label'
 export {default as LabelGroup} from './LabelGroup'
 export type {LabelGroupProps} from './LabelGroup'
 export {default as Link} from './Link'
